requirements.txt
annotated-types==0.7.0
asgiref==3.8.1
cfgv==3.4.0
coverage==7.6.4
distlib==0.3.9
Django==5.1.5
djangorestframework==3.15.2
filelock==3.16.1
gunicorn==23.0.0
identify==2.6.1
nodeenv==1.9.1
platformdirs==4.3.6
pydantic==2.10.1
pydantic_core==2.27.1
python-dotenv==1.0.1
PyYAML==6.0.2
ruff==0.7.1
sqlparse==0.5.1
typing_extensions==4.12.2
virtualenv==20.27.0
django-cors-headers==4.7.0
cryptography==45.0.3
PyJWT==2.10.1
requests==2.32.3
email-validator==2.2.0
drf-spectacular==0.28.0
psycopg2-binary==2.9.9
